100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,7 +6,7 @@
  * Return: Always 0 (Success)
  */
 
-long int main(void)
+int main(void)
 
 {
